% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Documentation.R
\name{createBicObject}
\alias{createBicObject}
\title{Return a BIC object, such as one you would find in the mclust package}
\usage{
createBicObject(modelNames, G, z = estep$z, data)
}
\arguments{
\item{modelNames}{names of possible models}

\item{G}{number of clusters}

\item{z}{A matrix whose [i,k]th entry is the probability that observation i in the test data belongs to the kth class}

\item{data}{the data you wish to use}
}
\value{
a BIC object (see mclustBIC in the mclust package)
}
\description{
Return a BIC object, such as one you would find in the mclust package
}
\examples{
#Load data
library(mclust)
data(banknote)

#Create new dataset with only continuous variables
bankdata <- banknote[,2:7]

#determine best model
model <- object$modelName
#extract model parameters
param <- object$parameters

#output from estep of EM algorithm
estepbank <- estep(model, bankdata, param)

#Create a BIC object, as seen in the mclust package's mclustBIC function
#Outputs BIC values for all models, but only for specified number of clusters
createBicObject(modelNames, G = object$G, z = estepbank$z, bankdata)
}
