% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Documentation.R
\name{imbc2}
\alias{imbc}
\title{A function that combines model-based clustering as well as your input to cluster your data}
\usage{
imbc(data, n = 1, G = NULL, query = "minimax",
  distanceMethod = "euclidean", iterationMax = 500)
}
\arguments{
\item{data}{the data you wish to use (must be continuous)}

\item{n}{the number of points you wish to be queried on at once}

\item{G}{number of clusters. The default allows Mclust to identify the number of clusters}

\item{query}{how you wish to be queried. This must be one of "minimax", "maxWithinClust" or "minBetweenClust", with a default of "minimax".}

\item{distanceMethod}{a method used to find distances between points. This must be one of "euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski", with a default of "euclidean."}

\item{iterationMax}{the maximum number of iterations you wish to see when converging mstep and estep}
}
\value{
An object providing the optimal (according to BIC) mixture model estimation.

The details of the output components are as follows:
\item{call}{The matched call}
\item{data}{The input data matrix}
\item{modelName}{A character string denoting the model at which the optimal BIC occurs}
\item{n}{The number of observations in the data}
\item{d}{The dimension of the data}
\item{G}{The optimal number of mixture components}
\item{BIC}{All BIC values}
\item{bic}{Optimal BIC value}
\item{loglik}{The log-likelihood corresponding to the optimal BIC}
\item{df}{The number of estimated parameters}
\item{hypvol}{The hypervolume parameter for the noise component if required, otherwise set to NULL (see hypvol).}
\item{parameters}{A list with the following components:}
\item{pro}{A vector whose kth component is the mixing proportion for the kth component of the mixture model. If missing, equal proportions are assumed}
\item{mean}{The mean for each component. If there is more than one component, this is a matrix whose kth column is the mean of the kth component of the mixture model.}
\item{variance}{A list of variance parameters for the model. The components of this list depend on the model specification. See the help file for mclustVariance for details.}
\item{z}{A matrix whose [i,k]th entry is the probability that observation i in the test data belongs to the kth class}
\item{classification}{The classification corresponding to z, i.e. map(z).}
\item{uncertainty}{The uncertainty associated with the classification.}
}
\description{
A function that combines model-based clustering as well as your input to cluster your data
}
\examples{
#Load data
library(mclust)
data(banknote)

#Create new dataset with only continuous variables
bankdata <- banknote[,2:7]

#Run imbc while querying user on onyl 1 data point at a time
#Use default querying algorithm (minimax)
output <- imbc(bankdata)

#query two points at once and using minimum between cluster distance as query method, and specifying 2 clusters
output2 <- imbc(bankdata, n = 2, G = 2, query = "minBetweenClust")

#gives vector of classification of each row
output2$classification

#classification probability matrix
#output2$z
}
